+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+[Unreleased]
+
 [0.1.1]
 ### Changed
 - Permission *view course* to *view all courses* changed
@@ -11,4 +13,5 @@
 ### Bugfix
 - corrected FeV paragraph
 
+[Unreleased]: https://github.com/prevplan/ausbilder.org/compare/v0.1.1...master
 [0.1.1]: https://github.com/prevplan/ausbilder.org/compare/v0.1...v0.1.1