+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Bugfix
 - replaces */* with *-* in the internal number
 - *Registration Link* linked
+- change company -> check activation
+- last company - save after creation
 
 ### Added
 - Honeypot for event search added